Switched to new drei-view, and updated implementation
