+++ src/components/canvas/Logo.jsx
@@ -1,0 +1,72 @@
+'use client'
+
+import { useState } from 'react';
+import { useThree } from '@react-three/fiber';
+import { OrbitControls, PerspectiveCamera, useGLTF } from '@react-three/drei';
+import { a, useSpring } from '@react-spring/three';
+import { useColor } from '@/utils/themes.js';
+
+export default function Logo({ ...props }) {
+  const [scale, setScale] = useState(Math.min(1, window.screen.width / 600));
+  const { nodes, materials } = useGLTF('/models/rp-logo.glb');
+
+  const { transparency } = useSpring({
+    config: { mass: 15, tension: 290, friction: 60, clamp: true },
+    to: {
+      transparency: 1
+    },
+    from: { transparency: 0 }
+  });
+
+  const { y } = useSpring({
+    config: { duration: 20000 },
+    loop: true,
+    from: { y: -0.065 },
+    to: { y: Math.PI * 2 - 0.065 }
+  });
+
+  const size = useThree((state) => state.size);
+
+  return (
+    <>
+      <group {...props} dispose={null}>
+        <a.mesh
+          geometry={nodes['RP_-_Logo001'].geometry}
+          rotation-y={y}
+          scale={size.width / 520}
+          material={materials.Abstract}
+          material-transparent
+          material-opacity={transparency}
+        />
+      </group>
+      <ambientLight intensity={0.1} />
+      {[
+        {
+          position: [-6, 1, 7],
+          color: useColor('--color-secondary'),
+          intensity: 110,
+          distance: 70,
+          decay: 2
+        },
+        {
+          position: [0, 1, 8],
+          color: useColor('--color-primary'),
+          intensity: 85,
+          distance: 80,
+          decay: 2
+        },
+        {
+          position: [6, 1, 7],
+          color: useColor('--color-secondary'),
+          intensity: 110,
+          distance: 70,
+          decay: 2
+        }
+      ].map((lightProps, index) => (
+        <pointLight key={index} {...lightProps} />
+      ))}
+    </>
+  );
+}
+
+useGLTF.preload('/models/rp-logo.glb');
