--- src/helpers/components/Three.jsx
@@ -1,7 +1,0 @@
-'use client'
-
-import { r3f } from '@/helpers/global'
-
-export const Three = ({ children }) => {
-  return <r3f.In>{children}</r3f.In>
-}
