--- src/helpers/global.js
@@ -1,3 +1,0 @@
-import tunnel from 'tunnel-rat'
-
-export const r3f = tunnel()
